+++ tools/build_pkg.R
@@ -2,7 +2,7 @@ pacman::p_load(styler, devtools, rhub)
 styler::style_pkg()
 devtools::document()
 devtools::spell_check()
-devtools::check()
+devtools::check(cran = TRUE)
 # rhub::rhub_doctor()
 # Note: "macos" (Intel macos-13) removed because rhub resolves it to
 # "macos-13-us-default" which is no longer available on GitHub Actions.